test(e2e): add tests for `@modern-js/e2e`
