+++ package.json
@@ -19,7 +19,7 @@
     "test": "npm run test:ut",
     "test:ut": "node --conditions=jsnext:source -r btsm node_modules/jest/bin/jest.js -c jest.config.js --maxWorkers=2 && npm run test:vitest",
     "test:e2e": "cd tests && npm run test",
-    "test:vitest": "pnpm --filter @modern-js/webpack-builder... run build && pnpm --filter @modern-js/webpack-builder run test",
+    "test:vitest": "pnpm --filter @modern-js/webpack-builder... run build && pnpm --filter @modern-js/webpack-builder --filter @modern-js/e2e run test",
     "fast-lint": "sh -x ./fast-lint.sh",
     "lint:package-json": "cd ./scripts/lint-package-json && pnpm start",
     "prepare": "pnpm run prepare --filter ./packages && husky install",
