+++ packages/toolkit/e2e/package.json
@@ -40,12 +40,17 @@
     "serve-static": "^1.14.1"
   },
   "devDependencies": {
+    "@scripts/vitest-config": "workspace:*",
     "@types/connect": "^3.4.35",
+    "@types/got": "^9.6.12",
     "@types/node": "^14",
     "@types/serve-static": "^1.13.10",
+    "@vitest/ui": "^0.21.1",
+    "got": "^11.8.3",
     "memfs": "^3.4.7",
     "nanoid": "3.3.4",
-    "typescript": "^4"
+    "typescript": "^4",
+    "vitest": "^0.21.1"
   },
   "wireit": {
     "build": {
