+++ packages/toolkit/e2e/tests/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  extends: ['@modern-js'],
+  parserOptions: {
+    project: require.resolve('./tsconfig.json'),
+  },
+};
